backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.10.0
pydantic-settings==2.6.0
pydantic[email]==2.10.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyMuPDF==1.24.13
google-genai==0.3.0
python-dotenv==1.0.1

